+++ default-plugins/status-bar/src/tip/data/floating_panes_mouse.rs
@@ -1,20 +1,7 @@
-use ansi_term::{unstyled_len, ANSIString, ANSIStrings, Style};
+use ansi_term::{ANSIString, Style};
 
-use crate::{action_key, style_key_with_modifier, LinePart};
+use crate::{action_key, ansi_strings, style_key_with_modifier, LinePart};
 use zellij_tile::prelude::{actions::Action, *};
-
-macro_rules! strings {
-    ($ANSIStrings:expr) => {{
-        let strings: &[ANSIString] = $ANSIStrings;
-
-        let ansi_strings = ANSIStrings(strings);
-
-        LinePart {
-            part: format!("{}", ansi_strings),
-            len: unstyled_len(&ansi_strings),
-        }
-    }};
-}
 
 pub fn floating_panes_mouse_full(help: &ModeInfo) -> LinePart {
     // Tip: Toggle floating panes with Ctrl + <p> + <w> and move them with keyboard or mouse
@@ -24,7 +11,7 @@ pub fn floating_panes_mouse_full(help: &ModeInfo) -> LinePart {
     ];
     bits.extend(add_keybinds(help));
     bits.push(Style::new().paint(" and move them with keyboard or mouse"));
-    strings!(&bits)
+    ansi_strings!(&bits)
 }
 
 pub fn floating_panes_mouse_medium(help: &ModeInfo) -> LinePart {
@@ -34,14 +21,14 @@ pub fn floating_panes_mouse_medium(help: &ModeInfo) -> LinePart {
         Style::new().paint("Toggle floating panes with "),
     ];
     bits.extend(add_keybinds(help));
-    strings!(&bits)
+    ansi_strings!(&bits)
 }
 
 pub fn floating_panes_mouse_short(help: &ModeInfo) -> LinePart {
     // Ctrl + <p> + <w> => floating panes
     let mut bits = add_keybinds(help);
     bits.push(Style::new().paint(" => floating panes"));
-    strings!(&bits)
+    ansi_strings!(&bits)
 }
 
 fn add_keybinds(help: &ModeInfo) -> Vec<ANSIString> {