+++ zellij-utils/src/plugin_api/action.rs
@@ -2,17 +2,18 @@ pub use super::generated_api::api::{
     action::{
         action::OptionalPayload, Action as ProtobufAction, ActionName as ProtobufActionName,
         DumpScreenPayload, EditFilePayload, GoToTabNamePayload, IdAndName,
-        LaunchOrFocusPluginPayload, MovePanePayload, NameAndValue as ProtobufNameAndValue,
-        NewFloatingPanePayload, NewPanePayload, NewPluginPanePayload, NewTiledPanePayload,
-        PaneIdAndShouldFloat, PluginConfiguration as ProtobufPluginConfiguration,
-        Position as ProtobufPosition, RunCommandAction as ProtobufRunCommandAction,
-        ScrollAtPayload, SearchDirection as ProtobufSearchDirection,
-        SearchOption as ProtobufSearchOption, SwitchToModePayload, WriteCharsPayload, WritePayload,
+        LaunchOrFocusPluginPayload, MovePanePayload, MoveTabDirection as ProtobufMoveTabDirection,
+        NameAndValue as ProtobufNameAndValue, NewFloatingPanePayload, NewPanePayload,
+        NewPluginPanePayload, NewTiledPanePayload, PaneIdAndShouldFloat,
+        PluginConfiguration as ProtobufPluginConfiguration, Position as ProtobufPosition,
+        RunCommandAction as ProtobufRunCommandAction, ScrollAtPayload,
+        SearchDirection as ProtobufSearchDirection, SearchOption as ProtobufSearchOption,
+        SwitchToModePayload, WriteCharsPayload, WritePayload,
     },
     input_mode::InputMode as ProtobufInputMode,
     resize::{Resize as ProtobufResize, ResizeDirection as ProtobufResizeDirection},
 };
-use crate::data::{Direction, InputMode, ResizeStrategy};
+use crate::data::{Direction, HorizontalDirection, InputMode, ResizeStrategy};
 use crate::errors::prelude::*;
 use crate::input::actions::Action;
 use crate::input::actions::{SearchDirection, SearchOption};
@@ -352,6 +353,16 @@ impl TryFrom<ProtobufAction> for Action {
             Some(ProtobufActionName::UndoRenameTab) => match protobuf_action.optional_payload {
                 Some(_) => Err("UndoRenameTab should not have a payload"),
                 None => Ok(Action::UndoRenameTab),
+            },
+            Some(ProtobufActionName::MoveTab) => match protobuf_action.optional_payload {
+                Some(OptionalPayload::MoveTabPayload(move_tab_payload)) => {
+                    let direction: HorizontalDirection =
+                        ProtobufMoveTabDirection::from_i32(move_tab_payload)
+                            .ok_or("Malformed move tab direction for Action::MoveTab")?
+                            .try_into()?;
+                    Ok(Action::MoveTab(direction))
+                },
+                _ => Err("Wrong payload for Action::MoveTab"),
             },
             Some(ProtobufActionName::Run) => match protobuf_action.optional_payload {
                 Some(OptionalPayload::RunPayload(run_command_action)) => {
@@ -984,6 +995,13 @@ impl TryFrom<Action> for ProtobufAction {
                 name: ProtobufActionName::UndoRenameTab as i32,
                 optional_payload: None,
             }),
+            Action::MoveTab(direction) => {
+                let direction: ProtobufMoveTabDirection = direction.try_into()?;
+                Ok(ProtobufAction {
+                    name: ProtobufActionName::MoveTab as i32,
+                    optional_payload: Some(OptionalPayload::MoveTabPayload(direction as i32)),
+                })
+            },
             Action::Run(run_command_action) => {
                 let run_command_action: ProtobufRunCommandAction = run_command_action.try_into()?;
                 Ok(ProtobufAction {
@@ -1289,6 +1307,28 @@ impl TryFrom<SearchDirection> for ProtobufSearchDirection {
         match search_direction {
             SearchDirection::Up => Ok(ProtobufSearchDirection::Up),
             SearchDirection::Down => Ok(ProtobufSearchDirection::Down),
+        }
+    }
+}
+
+impl TryFrom<ProtobufMoveTabDirection> for HorizontalDirection {
+    type Error = &'static str;
+    fn try_from(
+        protobuf_move_tab_direction: ProtobufMoveTabDirection,
+    ) -> Result<Self, &'static str> {
+        match protobuf_move_tab_direction {
+            ProtobufMoveTabDirection::Left => Ok(HorizontalDirection::Left),
+            ProtobufMoveTabDirection::Right => Ok(HorizontalDirection::Right),
+        }
+    }
+}
+
+impl TryFrom<HorizontalDirection> for ProtobufMoveTabDirection {
+    type Error = &'static str;
+    fn try_from(horizontal_direction: HorizontalDirection) -> Result<Self, &'static str> {
+        match horizontal_direction {
+            HorizontalDirection::Left => Ok(ProtobufMoveTabDirection::Left),
+            HorizontalDirection::Right => Ok(ProtobufMoveTabDirection::Right),
         }
     }
 }